requirements.txt
matplotlib>=3.6
numpy>=1.23
pyserial>=3.5
scikit-learn>=1.3
joblib>=1.3
PyYAML>=6.0
